refactor(session): extract tool runtime and prompt sanitizer
